+++ Modules/Intent.Modules.Ardalis.Repositories/Templates/TemplateExtensions.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Intent.Modelers.Domain.Api;
 using Intent.Modules.Ardalis.Repositories.Templates.ReadRepositoryInterface;
 using Intent.Modules.Common.Templates;
 using Intent.RoslynWeaver.Attributes;
@@ -11,14 +12,12 @@ namespace Intent.Modules.Ardalis.Repositories.Templates
 {
     public static class TemplateExtensions
     {
-
-        public static string GetReadRepositoryInterfaceName<T>(this IntentTemplateBase<T> template)
-where T : Intent.Modelers.Domain.Api.ClassModel
+        public static string GetReadRepositoryInterfaceName<T>(this IIntentTemplate<T> template) where T : ClassModel
         {
             return template.GetTypeName(ReadRepositoryInterfaceTemplate.TemplateId, template.Model);
         }
 
-        public static string GetReadRepositoryInterfaceName(this IntentTemplateBase template, Intent.Modelers.Domain.Api.ClassModel model)
+        public static string GetReadRepositoryInterfaceName(this IIntentTemplate template, ClassModel model)
         {
             return template.GetTypeName(ReadRepositoryInterfaceTemplate.TemplateId, model);
         }