+++ Tests/Ardalis/Ardalis.Application/Clients/DeleteClient/DeleteClientCommand.cs
@@ -1,0 +1,20 @@
+using System;
+using Ardalis.Application.Common.Interfaces;
+using Intent.RoslynWeaver.Attributes;
+using MediatR;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.CommandModels", Version = "1.0")]
+
+namespace Ardalis.Application.Clients.DeleteClient
+{
+    public class DeleteClientCommand : IRequest, ICommand
+    {
+        public DeleteClientCommand(Guid id)
+        {
+            Id = id;
+        }
+
+        public Guid Id { get; set; }
+    }
+}