+++ Tests/Ardalis/Ardalis.Application/Clients/UpdateClient/UpdateClientCommandValidator.cs
@@ -1,0 +1,24 @@
+using FluentValidation;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.FluentValidation.CommandValidator", Version = "2.0")]
+
+namespace Ardalis.Application.Clients.UpdateClient
+{
+    [IntentManaged(Mode.Fully, Body = Mode.Merge)]
+    public class UpdateClientCommandValidator : AbstractValidator<UpdateClientCommand>
+    {
+        [IntentManaged(Mode.Merge)]
+        public UpdateClientCommandValidator()
+        {
+            ConfigureValidationRules();
+        }
+
+        private void ConfigureValidationRules()
+        {
+            RuleFor(v => v.Name)
+                .NotNull();
+        }
+    }
+}