+++ Tests/Ardalis/Ardalis.Domain/Common/Exceptions/NotFoundException.cs
@@ -1,0 +1,15 @@
+using System;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Entities.NotFoundException", Version = "1.0")]
+
+namespace Ardalis.Domain.Common.Exceptions
+{
+    public class NotFoundException : Exception
+    {
+        public NotFoundException(string message) : base(message)
+        {
+        }
+    }
+}