+++ Tests/Ardalis/Ardalis.Infrastructure/DependencyInjection.cs
@@ -1,0 +1,29 @@
+using Ardalis.Domain.Common.Interfaces;
+using Ardalis.Domain.Repositories;
+using Ardalis.Infrastructure.Persistence;
+using Ardalis.Infrastructure.Repositories;
+using Intent.RoslynWeaver.Attributes;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Infrastructure.DependencyInjection.DependencyInjection", Version = "1.0")]
+
+namespace Ardalis.Infrastructure
+{
+    public static class DependencyInjection
+    {
+        public static IServiceCollection AddInfrastructure(this IServiceCollection services, IConfiguration configuration)
+        {
+            services.AddDbContext<ApplicationDbContext>((sp, options) =>
+            {
+                options.UseInMemoryDatabase("DefaultConnection");
+                options.UseLazyLoadingProxies();
+            });
+            services.AddScoped<IUnitOfWork>(provider => provider.GetRequiredService<ApplicationDbContext>());
+            services.AddTransient<IClientRepository, ClientRepository>();
+            return services;
+        }
+    }
+}