+++ Modules/Intent.Modules.Ardalis.GuardClauses/release-notes.md
@@ -1,3 +1,7 @@
+### Version 4.1.0
+
+- Improvement: Updated module project to .NET 8.
+
 ### Version 4.0.2
 
 - Updated dependencies and supported client versions to prevent warnings when used with Intent Architect 4.x.